# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SIU-2 Backend API
  version: 1.0.0
  description: Documentación OpenAPI 3.0 generada exclusivamente a partir de los routers definidos en routes/.
servers:
  - url: http://localhost:5000/api
    description: Servidor local
...
security:
  - CookieAuth: []
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: Cookie con el JWT de sesión.
  schemas:
    ErrorItem:
      type: object
      required:
        - code
        - message
        - level
      properties:
        code:
          type: string
        message:
          type: string
        level:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
...
    Materia:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
    MateriaCreateRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
    MateriaUpdateRequest:
      type: object
      required:
        - nombre
        - codigo
      properties:
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
...
    Evaluacion:
      type: object
      properties:
        id:
          type: integer
        curso_id:
          type: integer
        tipo_evaluacion_id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
          nullable: true
        fecha:
          type: string
          format: date
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
        es_grupal:
          type: boolean
        tipo_evaluacion:
          type: string
    EvaluacionCreateRequest:
      type: object
      required:
        - curso_id
        - tipo_evaluacion_id
        - titulo
        - fecha
      properties:
        curso_id:
          type: integer
        tipo_evaluacion_id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
          nullable: true
        fecha:
          type: string
          format: date
    EvaluacionUpdateRequest:
      type: object
      required:
        - curso_id
        - tipo_evaluacion_id
        - titulo
        - fecha
        - activo
      properties:
        curso_id:
          type: integer
        tipo_evaluacion_id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
          nullable: true
        fecha:
          type: string
          format: date
        activo:
          type: boolean
    Equipo:
      type: object
      properties:
        id:
          type: integer
        curso_id:
          type: integer
        evaluacion_id:
          type: integer
        nombre:
          type: string
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
    EquipoCreateRequest:
      type: object
      required:
        - curso_id
        - evaluacion_id
        - nombre
      properties:
        curso_id:
          type: integer
        evaluacion_id:
          type: integer
        nombre:
          type: string
    EquipoUpdateRequest:
      type: object
      required:
        - curso_id
        - evaluacion_id
        - nombre
        - activo
      properties:
        curso_id:
          type: integer
        evaluacion_id:
          type: integer
        nombre:
          type: string
        activo:
          type: boolean
...
    EquipoIntegrante:
      type: object
      properties:
        equipo_id:
          type: integer
        alumno_id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        activo:
          type: boolean
    EquipoIntegranteCreateRequest:
      type: object
      required:
        - equipo_id
        - alumno_id
      properties:
        equipo_id:
          type: integer
        alumno_id:
          type: integer
...
  responses:
    BadRequest:
      description: Solicitud inválida o datos de entrada no válidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Falta la cookie de autenticación, el token es inválido o expiró.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no posee el rol requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: El recurso ya existe o viola una restricción de unicidad.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /materias:
    get:
      tags: [Materias]
      summary: Listar materias
      parameters:
        - name: nombre
          in: query
          schema:
            type: string
        - name: codigo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de materias.
          content:
            application/json:
              schema:
                type: object
                required: [materias, total, page, page_size]
                properties:
                  materias:
                    type: array
                    items:
                      $ref: '#/components/schemas/Materia'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Materias]
      summary: Crear una materia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaCreateRequest'
      responses:
        '201':
          description: Materia creada correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, materia]
                properties:
                  message:
                    type: string
                  materia:
                    $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /materias/{materia_id}:
    parameters:
      - name: materia_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Materias]
      summary: Obtener una materia por ID
      responses:
        '200':
          description: Materia encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Materias]
      summary: Reemplazar una materia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaUpdateRequest'
      responses:
        '200':
          description: Materia actualizada correctamente.
          content:
            application/json:
              schema:
                type: object
                required: [message, materia]
                properties:
                  message:
                    type: string
                  materia:
                    $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Materias]
      summary: Eliminar una materia
      responses:
        '204':
          description: Materia eliminada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /evaluaciones:
    get:
      tags: [Evaluaciones]
      summary: Listar evaluaciones
      parameters:
        - name: curso_id
          in: query
          schema:
            type: integer
        - name: tipo_evaluacion_id
          in: query
          schema:
            type: integer
        - name: titulo
          in: query
          schema:
            type: string
        - name: fecha
          in: query
          schema:
            type: string
            format: date
        - name: activo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de evaluaciones.
          content:
            application/json:
              schema:
                type: object
                required: [evaluaciones, total]
                properties:
                  evaluaciones:
                    type: array
                    items:
                      $ref: '#/components/schemas/Evaluacion'
                  total:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Evaluaciones]
      summary: Crear una evaluación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluacionCreateRequest'
      responses:
        '201':
          description: Evaluación creada correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, evaluacion]
                properties:
                  message:
                    type: string
                  evaluacion:
                    $ref: '#/components/schemas/Evaluacion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /evaluaciones/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Evaluaciones]
      summary: Obtener una evaluación por ID
      responses:
        '200':
          description: Evaluación encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluacion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Evaluaciones]
      summary: Reemplazar una evaluación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluacionUpdateRequest'
      responses:
        '204':
          description: Evaluación actualizada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Evaluaciones]
      summary: Eliminar una evaluación
      responses:
        '204':
          description: Evaluación eliminada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /equipos:
    get:
      tags: [Equipos]
      summary: Listar equipos
      parameters:
        - name: curso_id
          in: query
          schema:
            type: integer
        - name: evaluacion_id
          in: query
          schema:
            type: integer
        - name: nombre
          in: query
          schema:
            type: string
        - name: activo
          in: query
          schema:
            type: string
            description: El código lo lee como texto y compara true/false.
      responses:
        '200':
          description: Lista de equipos.
          content:
            application/json:
              schema:
                type: object
                required: [equipos, total]
                properties:
                  equipos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Equipo'
                  total:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Equipos]
      summary: Crear un equipo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipoCreateRequest'
      responses:
        '201':
          description: Equipo creado correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, equipo]
                properties:
                  message:
                    type: string
                  equipo:
                    $ref: '#/components/schemas/Equipo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /equipos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Equipos]
      summary: Obtener un equipo por ID
      responses:
        '200':
          description: Equipo encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Equipos]
      summary: Reemplazar un equipo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipoUpdateRequest'
      responses:
        '204':
          description: Equipo actualizado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: El equipo no existe.
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Equipos]
      summary: Eliminar un equipo
      responses:
        '204':
          description: Equipo eliminado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /equipo_integrantes:
    get:
      tags: [EquipoIntegrantes]
      summary: Listar integrantes de equipos
      parameters:
        - name: equipo_id
          in: query
          schema:
            type: integer
        - name: alumno_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de integrantes.
          content:
            application/json:
              schema:
                type: object
                required: [integrantes, total]
                properties:
                  integrantes:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipoIntegrante'
                  total:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [EquipoIntegrantes]
      summary: Agregar integrante a un equipo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipoIntegranteCreateRequest'
      responses:
        '201':
          description: Integrante agregado correctamente.
          content:
            application/json:
              schema:
                type: object
                required: [message, integrante]
                properties:
                  message:
                    type: string
                  integrante:
                    $ref: '#/components/schemas/EquipoIntegrante'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /equipo_integrantes/{equipo_id}/{alumno_id}:
    parameters:
      - name: equipo_id
        in: path
        required: true
        schema:
          type: integer
      - name: alumno_id
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags: [EquipoIntegrantes]
      summary: Eliminar integrante de un equipo
      responses:
        '204':
          description: Integrante eliminado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...